requirements.txt
mysql-connector-python==9.5.0
python-dotenv==1.2.1
Faker>=30.0.0
tqdm>=4.67.0